ai-pipeline/requirements.txt
# FastAPI & Server
fastapi>=0.109
uvicorn[standard]>=0.27
python-multipart>=0.0.6

# ===================
# LLM & AI
# ===================
# OpenAI & Anthropic (API-based)
openai>=1.10
anthropic>=0.14

# Ollama (Local LLM - NO API KEY REQUIRED)
ollama>=0.1.6

# LangChain Ecosystem
langchain>=0.1.4
langchain-core>=0.1.17
langchain-community>=0.0.17
langchain-openai>=0.0.5
langchain-anthropic>=0.1
langgraph>=0.0.20

# LangSmith (Monitoring & Tracing)
langsmith>=0.0.83

# ===================
# Vector Databases
# ===================
# ChromaDB (Local vector DB)
chromadb>=0.4.22

# Pinecone (Cloud alternative)
pinecone-client>=3.0

# Weaviate (Alternative)
weaviate-client>=4.4

# FAISS (CPU-based similarity search)
faiss-cpu>=1.7

# ===================
# Embeddings
# ===================
sentence-transformers>=2.3
numpy>=1.26

# ===================
# Document Processing
# ===================
pdfplumber>=0.10
python-docx>=1.1
PyPDF2>=3.0
pytesseract>=0.3
pdf2image>=1.17
unstructured>=0.12

# ===================
# Video & Audio Processing
# ===================
ffmpeg-python>=0.2
moviepy>=1.0
pydub>=0.25
opencv-python-headless>=4.9

# Text-to-Speech
elevenlabs>=0.2
gtts>=2.5
pyttsx3>=2.90

# Speech-to-Text
openai-whisper>=20231117
faster-whisper>=0.10

# ===================
# Avatar Video Generation APIs
# ===================
httpx>=0.26
aiohttp>=3.9

# ===================
# Supabase (Database & Auth)
# ===================
supabase>=2.3

# ===================
# Message Queue & Caching
# ===================
kafka-python>=2.0
aiokafka>=0.10
redis>=5.0
aioredis>=2.0

# ===================
# Storage
# ===================
boto3>=1.34
minio>=7.2

# ===================
# Hallucination Detection
# ===================
# SelfCheckGPT, Guardrails
guardrails-ai>=0.4

# ===================
# Utils
# ===================
python-dotenv>=1.0
pyyaml>=6.0
Pillow>=10.2
python-magic>=0.4
tenacity>=8.2  # Retry logic
tiktoken>=0.5  # Token counting

# ===================
# Monitoring & Logging
# ===================
prometheus-client>=0.19
structlog>=24.1
sentry-sdk>=1.39

# ===================
# Testing
# ===================
pytest>=7.4
pytest-asyncio>=0.23
httpx>=0.26

# ===================
# Code Quality
# ===================
black>=24.1
isort>=5.13
